fix(transmission): emit integer average speeds in torrent properties
go-qbittorrent decodes dl_speed_avg/up_speed_avg as integers; the
previous float64 division serialized fractional values and made the
client reject the whole properties response. Use integer division and
teach the fake daemon to serve secondsDownloading/secondsSeeding and
isPrivate so the test asserts real values.

+++ internal/transmission/bridge.go
@@ -675,10 +675,12 @@ func (b *Bridge) handleTorrentProperties(ctx context.Context, req *http.Request)
 
 	qt := qbitTorrentFrom(t)
 	timeActive := qt.TimeActive
-	dlAvg, upAvg := 0.0, 0.0
+	// qBittorrent exposes average speeds as integers; fractional values make
+	// go-qbittorrent reject the entire properties response.
+	dlAvg, upAvg := int64(0), int64(0)
 	if timeActive > 0 {
-		dlAvg = float64(t.DownloadedEver) / float64(timeActive)
-		upAvg = float64(t.UploadedEver) / float64(timeActive)
+		dlAvg = t.DownloadedEver / timeActive
+		upAvg = t.UploadedEver / timeActive
 	}
 
 	props := map[string]any{
